waterbody dataframe updates by Sean Horvath
